Major Update
1. Finally implemented End-Game Algorithm and some more enhancements, such as dropping choked/not having our pieces peers in scheduler (It would not finish downloading, usually at the end of torrent)

2. BitField algorithms updates for more complex request piece algorithms later on

3. Integer Overflow Issue with large files (more than Int32.MaxValue (It was crashing during WritePiece and CreateFiles)

+++ TorSwarm/PartFile.cs
@@ -115,10 +115,10 @@ namespace SuRGeoNix.TorSwarm
 
                 int chunkPos    = mapIdToChunkId[chunkId];
                 int chunkPos2   = chunkPos;
-                int pos         = 0;
+                long pos         = 0;
                 if ( firstChunkSize != 0 && chunkPos > firstPos)  { pos += firstChunkSize;    chunkPos2--; }
                 if ( lastChunkSize  != 0 && chunkPos > lastPos )  { pos += lastChunkSize;     chunkPos2--; }
-                pos += ChunkSize * chunkPos2;
+                pos += (long)ChunkSize * chunkPos2;
 
                 byte[] data = new byte[len];
                 fileStream.Seek(pos, SeekOrigin.Begin);
